requirements.txt
asgiref==3.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1.1
djangorestframework==3.13.1
idna==3.4
python-decouple==3.8
pytz==2022.2.1
requests==2.31.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==2.0.4
